requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.*
Flask==2.*
getenv==0.*
idna==3.*
itsdangerous==2.*
Jinja2==3.*
MarkupSafe==2.*
PyJWT==2.*
pymongo==4.*
python-dotenv==0.20.0
requests==2.*
urllib3==1.26.9
Werkzeug==2.*
